gomod: extract utils.py and test_utils.py
Move _clean_go_modcache and _go_exec_env into utils.py.
Extract test_go_exec_env into test_utils.py.
