+++ tests/unit/package_managers/gomod/test_main.py
@@ -52,7 +52,6 @@ from hermeto.core.package_managers.gomod.main import (
     _get_go_sum_files,
     _get_proxy_for_module,
     _get_repository_name,
-    _go_exec_env,
     _go_list_deps,
     _parse_go_sum,
     _parse_local_modules,
@@ -1933,41 +1932,6 @@ def test_parse_packages(
     # _parse_packages calls _go_list_deps always with the './...' pattern
     assert all("./..." in call.args[0] for call in calls)
     assert set(pkgs) == expected
-
-
-_ENV_VARS_BASE_INIT = {v: "/some/path" for v in ("PATH", "HOME", "NETRC")}
-
-
-@pytest.mark.parametrize(
-    "env, extra_env, expected",
-    [
-        pytest.param(_ENV_VARS_BASE_INIT, None, _ENV_VARS_BASE_INIT, id="vars_inherited"),
-        pytest.param(
-            {},
-            None,
-            {"PATH": "", "HOME": "/mocked/home", "NETRC": ""},
-            id="vars_defaults",
-        ),
-        pytest.param(
-            _ENV_VARS_BASE_INIT,
-            {"GOPATH": "/tmp/go"},
-            _ENV_VARS_BASE_INIT | {"GOPATH": "/tmp/go"},
-            id="with_extra_env",
-        ),
-    ],
-)
-@mock.patch("pathlib.Path.home", return_value=Path("/mocked/home"))
-def test_go_exec_env(
-    mock_home: mock.Mock,
-    monkeypatch: pytest.MonkeyPatch,
-    env: dict[str, str],
-    extra_env: dict[str, str] | None,
-    expected: dict[str, str],
-) -> None:
-    monkeypatch.setattr(os, "environ", env)
-
-    actual = _go_exec_env() if extra_env is None else _go_exec_env(**extra_env)
-    assert actual == expected
 
 
 @pytest.mark.parametrize(
